refactor(card-table): split grid_view.py into view + layout + images + render + interaction (#798)
DeckGridView was a single 1014-LOC class banding ~9 responsibilities already
separated by section comments with low cross-coupling. Decompose it in the
package's lighter view-collaborator style — thin mixins the concrete view
inherits, NOT a new view-level protocol.py — matching the convention marquee.py
/ scrolling.py / card_render.py already establish for these scrolled-window
views.

This is a behaviour-preserving move: code is relocated verbatim (only imports
adjusted), no runtime behaviour, public API or wx event-wiring semantics change.

Split:
- grid_view.py (~278): slimmed DeckGridView — __init__, all instance-state
  declarations, event binding, the public CardTablePanel-facing API
  (set_cards/set_selected/get_selected_*/focus_card/count_loaded_images/
  refresh_image) and _on_paint orchestration. Inherits the four mixins with
  wx.ScrolledWindow LAST in the MRO; every attribute the mixins read via self
  is still declared here.
- grid_layout.py (~121): pure geometry + the shared cell/grid module constants
  (_CELL_*, _ACTION_*, _DEFAULT_COLUMNS, _MAX_CANVAS_PX) — _recompute_layout,
  _card_rect, _hit_test, _to_logical, _ensure_visible, _action_button_rects,
  _drop_index_at. Lowest-risk band; other grid modules import its constants.
- grid_images.py (~216): async decode pipeline — module-level shared decode
  pool + atexit non-blocking shutdown, _prefetch_images/_start_image_load/
  _image_worker/_image_loaded and the placeholder-template builders. The
  off-thread -> wx.CallAfter -> _image_loaded liveness contract (the `if not
  self` validity guard + _load_gen staleness check) is kept intact.
- grid_render.py (~215): cached canvas + drawing primitives —
  _invalidate_canvas/_ensure_canvas/_patch_card_on_canvas, _visible_card_indices,
  _draw_overlays/_draw_card_static/_draw_card/_draw_qty/_draw_actions.
- grid_interaction.py (~333): event handlers + selection state machine + marquee
  glue + drag-to-reorder (incl. the ghost/drop-indicator overlays drawn in
  _on_paint). The _suppress_set_selected echo-guard and the marquee
  multi-select round-trip are preserved verbatim.
